+++ Faker_top_level_domain.v
@@ -1,0 +1,5 @@
+module faker
+
+pub fn (mut this Faker) top_level_domain() string {
+    return this.random_element(TopLevelDomain.cases()).str()
+}