+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Currency symbol fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 - Currency code fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 - User Agent (browser) fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
+- Top level domain fake data ([#3](https://github.com/khalyomede/faker/issues/3)).
 
 ### Fixed